headers: isolate OpenGL compatibility fallbacks
Move the complete historical fallback block out of Inventor/system/gl.h without changing the public macro contract. Keep the compatibility declarations on the public platform-header path so older consumer toolchains retain every definition Coin previously supplied.
